+++ static/js/server_events.js
@@ -139,7 +139,7 @@ function dispatch_normal_event(event) {
 
     case 'realm_filters':
         page_params.realm_filters = event.realm_filters;
-        echo.set_realm_filters(page_params.realm_filters);
+        markdown.set_realm_filters(page_params.realm_filters);
         settings_filters.populate_filters(page_params.realm_filters);
         break;
